undo the changes for logout button due to some complexities
